update BaseStyle and AvatarGroupView
