+++ sample/src/main/res/layout/activity_channel.xml
@@ -3,7 +3,9 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:stream="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools">
+
     <data>
+
         <variable
             name="viewModel"
             type="com.getstream.sdk.chat.viewmodel.ChannelViewModel" />
@@ -16,47 +18,37 @@
 
         <com.getstream.sdk.chat.view.ChannelHeaderToolbar
             android:id="@+id/channelHeader"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:background="#2F3036"
             app:layout_constraintEnd_toStartOf="@+id/messageList"
-            android:background="#2F3036" android:layout_height="wrap_content"
-            android:layout_width="match_parent" />
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
 
-        <com.getstream.sdk.chat.view.MessageListView android:id="@+id/messageList"
+        <com.getstream.sdk.chat.view.MessageListView
+            android:id="@+id/messageList"
             android:layout_width="match_parent"
             android:layout_height="0dp"
             android:layout_marginBottom="10dp"
-            app:layout_constraintBottom_toTopOf="@+id/message_input"
-            app:layout_constraintTop_toBottomOf="@+id/channelHeader"
-            app:layout_constraintStart_toEndOf="@+id/channelHeader"
-            app:layout_constraintEnd_toStartOf="@+id/message_input"
             android:background="#F3F5F8"
-            stream:messageTextColorMine="#FFFFFF"
+            app:layout_constraintBottom_toTopOf="@+id/message_input"
+            app:layout_constraintEnd_toStartOf="@+id/message_input"
+            app:layout_constraintStart_toEndOf="@+id/channelHeader"
+            app:layout_constraintTop_toBottomOf="@+id/channelHeader"
             stream:messageBubbleDrawableMine="@drawable/message_bubble_mine"
-            stream:messageTextColorTheirs="#1BB76E"
             stream:messageBubbleDrawableTheirs="@drawable/message_bubble_theirs"
+            stream:messageTextColorMine="#FFFFFF"
+            stream:messageTextColorTheirs="#1BB76E"
             tools:listitem="@layout/list_item_message" />
 
+
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="unused"
-            app:layout_constraintLeft_toLeftOf="parent"
-
-            app:layout_constraintTop_toTopOf="parent"
-            android:padding="6dp"
             android:layout_marginTop="16dp"
-            />
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
+            android:padding="6dp"
             app:layout_constraintRight_toRightOf="parent"
-
-            app:layout_constraintTop_toTopOf="parent"
-            android:padding="6dp"
-            android:layout_marginTop="16dp"
-            />
+            app:layout_constraintTop_toTopOf="parent" />
 
         <ProgressBar
             android:layout_width="wrap_content"
@@ -77,19 +69,16 @@
 
         <com.getstream.sdk.chat.view.MessageInputView
             android:id="@+id/message_input"
-            android:layout_height="wrap_content"
             android:layout_width="match_parent"
-            android:background="@color/chat_theme"
-            android:layout_marginBottom="0dp"
+            android:layout_height="wrap_content"
             android:layout_marginTop="32dp"
+            android:layout_marginBottom="0dp"
+            android:background="@color/chat_theme"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toEndOf="@+id/messageList"
-            app:layout_constraintEnd_toEndOf="parent"
-            stream:inputHint="what a fun day"
-
-            />
-
+            stream:inputHint="what a fun day"/>
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
